feat: Return token from auth endpoints
Why:

I didn't want any socket io instance to be able to just connect to the
server socket. For a socket to  connect to the server, they need to be a
registered user, and to prove so, the idea was to send the token
associated with their account upon requesting the connection by adding
to the `auth` object in socket.io's `opts`.

Downsides:

I'm not sure if it's wise to send the token that was originally (and
still being) used as a cookie to keep the users logged in. The cookie
was saved as an httpOnly cookie so that it can't be access via
javascript. This solution is temporary, will keep it till I come up with
a better idea for authenticating sockets, or till it proves to be safe.