feat: Receive cookies upon initial conn. request
Setting credentials to true allows the server to include cookies,
HTTP authentication, and client-side SSL/TLS certificates in the
cross-origin request.

This is important since the client is trying to make cross-origin
requests that require user authentication, specifically the "token" cookie.

Now, the server doesn't have to send the token in the auth
endpoints, since that was their only use, inital establishment of socket
connections.