feat: Redo authentication client-side
Authentication client-side was a disaster, it had to be redone from
scratch.

How it was:

AuthContext provided the current user info and basic async auth functions
like login and logout etc. but they were used in a wrong way.

I found myself using these auth functions with useAsync to show loading
components while the server takes the time to respond. If a user went to
a page that requires auth and the response for the /auth/me request is
not returned yet, they will be redirected to / even though they are
logged in but the server didn't respond fast enough.

How it is now:

The loading state is now global in the AuthContext itself, so that all
components can know when ANYTHING is loading and behave accordingly.